[ANDROSDK-1046] Move packages and create DI structure
