+++ core/src/main/java/org/hisp/dhis/android/core/settings/android/DataSyncPeriod.java
@@ -26,36 +26,12 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.configuration;
+package org.hisp.dhis.android.core.settings.android;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-import com.google.auto.value.AutoValue;
-
-import java.util.Map;
-
-@AutoValue
-@JsonDeserialize(builder = AutoValue_DataSetSettings.Builder.class)
-public abstract class DataSetSettings {
-
-    @JsonProperty()
-    public abstract DataSetSettingsItem globalSettings();
-
-    @JsonProperty()
-    public abstract Map<String, DataSetSettingsItem> specificSettings();
-
-    public static Builder builder() {
-        return new AutoValue_DataSetSettings.Builder();
-    }
-
-    @AutoValue.Builder
-    @JsonPOJOBuilder(withPrefix = "")
-    public abstract static class Builder {
-        public abstract Builder globalSettings(DataSetSettingsItem globalSettings);
-
-        public abstract Builder specificSettings(Map<String, DataSetSettingsItem> specificSettings);
-
-        public abstract DataSetSettings build();
-    }
+public enum DataSyncPeriod {
+    EVERY_30_MIN,
+    EVERY_HOUR,
+    EVERY_6_HOURS,
+    EVERY_12_HOURS,
+    EVERY_24_HOURS
 }
