+++ core/src/main/java/org/hisp/dhis/android/core/settings/android/LimitScope.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.configuration;
+package org.hisp.dhis.android.core.settings.android;
 
 public enum LimitScope {
     GLOBAL,
