+++ core/src/sharedTest/resources/settings/android/dataset_settings.json
